refactor(workflow): extract DerivedPlanTaskContractBuilder
- move buildDerivedTaskContract/title/description and isSameDerivedPlanTask
  out of WorkflowService into a dedicated pure builder class
- workflowService.ts shrinks from 1189 to 1064 lines
- drops unused local imports (PlanPhase, PlanStep, getPhaseDefinition)
