+++ src/services/workflow/workflowService.ts
@@ -10,8 +10,8 @@ import { exec as execCallback } from 'child_process';
 import { promisify } from 'util';
 import { CollaborationSession, CollaborationManager } from '../../workflow/collaboration';
 import { GateCheckResult } from '../../workflow/gates';
-import { getPhaseDefinition, PHASE_NAMES_PT } from '../../workflow/phases';
-import { createPlanLinker, type LinkedPlan, type PlanPhase, type PlanStep } from '../../workflow/plans';
+import { PHASE_NAMES_PT } from '../../workflow/phases';
+import { createPlanLinker, type LinkedPlan } from '../../workflow/plans';
 import { ROLE_DISPLAY_NAMES } from '../../workflow/roles';
 import { getScaleName, getScaleFromName } from '../../workflow/scaling';
 import { PrevcOrchestrator, type WorkflowSummary } from '../../workflow/orchestrator';
@@ -29,6 +29,7 @@ import type { CollaborationSynthesis } from '../../workflow/types';
 import {
   FileCollaborationStore,
 } from './fileCollaborationStore';
+import { DerivedPlanTaskContractBuilder } from './derivedPlanTaskContractBuilder';
 import {
   HarnessRuntimeStateService,
   HarnessSensorCatalogService,
@@ -46,14 +47,6 @@ import {
 } from '../harness';
 
 const exec = promisify(execCallback);
-
-function uniqueStrings(values: Array<string | undefined | null>): string[] {
-  return Array.from(new Set(
-    values
-      .map((value) => typeof value === 'string' ? value.trim() : '')
-      .filter((value) => value.length > 0)
-  ));
-}
 
 export interface WorkflowHarnessStatus {
   binding: WorkflowHarnessBinding;
@@ -84,18 +77,6 @@ export class HarnessWorkflowBlockedError extends Error {
     super(message);
     this.name = 'HarnessWorkflowBlockedError';
   }
-}
-
-interface DerivedTaskContractInput {
-  title: string;
-  description?: string;
-  owner?: string;
-  inputs: string[];
-  expectedOutputs: string[];
-  acceptanceCriteria: string[];
-  requiredSensors: string[];
-  requiredArtifacts: string[];
-  metadata: Record<string, unknown>;
 }
 
 interface WorkflowPlanLinkResult {
@@ -149,6 +130,7 @@ export class WorkflowService {
   private taskContractsService: HarnessTaskContractsService;
   private policyService: HarnessPolicyService;
   private workflowStateService: HarnessWorkflowStateService;
+  private derivedTaskBuilder: DerivedPlanTaskContractBuilder;
   private deps: WorkflowServiceDependencies;
 
   constructor(
@@ -178,6 +160,7 @@ export class WorkflowService {
     this.collaborationManager = new CollaborationManager(
       new FileCollaborationStore(this.contextPath)
     );
+    this.derivedTaskBuilder = new DerivedPlanTaskContractBuilder();
     this.deps = deps;
     this.registerDefaultSensors();
   }
@@ -545,8 +528,8 @@ export class WorkflowService {
       return null;
     }
 
-    const contractInput = this.buildDerivedTaskContract(linkedPlan, params.phase);
-    if (this.isSameDerivedPlanTask(params.activeTask, linkedPlan.ref.slug, params.phase)) {
+    const contractInput = this.derivedTaskBuilder.build(linkedPlan, params.phase);
+    if (this.derivedTaskBuilder.isSameDerivedTask(params.activeTask, linkedPlan.ref.slug, params.phase)) {
       return this.taskContractsService.updateTaskContract(params.activeTask.id, {
         title: contractInput.title,
         description: contractInput.description,
@@ -580,25 +563,12 @@ export class WorkflowService {
       return null;
     }
 
-    const contractInput = this.buildDerivedTaskContract(linkedPlan, params.phase);
+    const contractInput = this.derivedTaskBuilder.build(linkedPlan, params.phase);
     return this.taskContractsService.createTaskContract({
       ...contractInput,
       sessionId: params.sessionId,
       status: 'ready',
     });
-  }
-
-  private isSameDerivedPlanTask(
-    activeTask: HarnessTaskContract | null | undefined,
-    planSlug: string,
-    phase: PrevcPhase
-  ): activeTask is HarnessTaskContract {
-    return Boolean(
-      activeTask &&
-      activeTask.metadata?.source === 'workflow.plan' &&
-      activeTask.metadata?.planSlug === planSlug &&
-      activeTask.metadata?.prevcPhase === phase
-    );
   }
 
   /**
@@ -1002,101 +972,6 @@ export class WorkflowService {
     }
 
     return linker.getLinkedPlan(resolvedPlanSlug);
-  }
-
-  private buildDerivedTaskContract(plan: LinkedPlan, phase: PrevcPhase): DerivedTaskContractInput {
-    const matchingPlanPhases = plan.phases.filter((planPhase) => planPhase.prevcPhase === phase);
-    const phaseDefinition = getPhaseDefinition(phase);
-    const flattenedSteps = matchingPlanPhases.flatMap((planPhase) => planPhase.steps);
-    const explicitOutputs = [
-      ...matchingPlanPhases.flatMap((planPhase) => planPhase.deliverables ?? []),
-      ...flattenedSteps.flatMap((step) => step.outputs || step.deliverables || []),
-    ];
-    const expectedOutputs = uniqueStrings(
-      explicitOutputs.length > 0
-        ? explicitOutputs
-        : phaseDefinition.outputs
-    );
-    const acceptanceCriteria = uniqueStrings(
-      flattenedSteps.length > 0
-        ? flattenedSteps.map((step) => step.description)
-        : matchingPlanPhases.length > 0
-          ? matchingPlanPhases.map((planPhase) => planPhase.summary || `Complete ${planPhase.name}`)
-          : [`Complete ${phaseDefinition.name}`]
-    );
-    const owners = uniqueStrings(
-      flattenedSteps
-        .map((step) => step.assignee)
-        .filter((assignee): assignee is string => typeof assignee === 'string' && assignee.trim().length > 0)
-    );
-    const owner = owners.length === 1 ? owners[0] : phaseDefinition.roles[0];
-    const title = this.buildDerivedTaskTitle(plan, phase, matchingPlanPhases);
-    const description = this.buildDerivedTaskDescription(plan, phase, matchingPlanPhases, flattenedSteps, expectedOutputs);
-
-    return {
-      title,
-      description,
-      owner,
-      inputs: plan.docs.map((doc) => `docs/${doc}`),
-      expectedOutputs,
-      acceptanceCriteria,
-      requiredSensors: [],
-      requiredArtifacts: [],
-      metadata: {
-        source: 'workflow.plan',
-        planSlug: plan.ref.slug,
-        prevcPhase: phase,
-        planPhaseIds: matchingPlanPhases.map((planPhase) => planPhase.id),
-        planPhaseNames: matchingPlanPhases.map((planPhase) => planPhase.name),
-        planPhaseSummaries: matchingPlanPhases.map((planPhase) => planPhase.summary).filter(Boolean),
-        derivedFrom: matchingPlanPhases.length > 0 ? 'linked-plan' : 'workflow-phase-defaults',
-      },
-    };
-  }
-
-  private buildDerivedTaskTitle(plan: LinkedPlan, phase: PrevcPhase, matchingPlanPhases: PlanPhase[]): string {
-    if (matchingPlanPhases.length === 1) {
-      return `${plan.ref.title}: ${matchingPlanPhases[0].name}`;
-    }
-
-    if (matchingPlanPhases.length > 1) {
-      return `${plan.ref.title}: ${getPhaseDefinition(phase).name} (${matchingPlanPhases.length} plan phases)`;
-    }
-
-    return `${plan.ref.title}: ${getPhaseDefinition(phase).name}`;
-  }
-
-  private buildDerivedTaskDescription(
-    plan: LinkedPlan,
-    phase: PrevcPhase,
-    matchingPlanPhases: PlanPhase[],
-    flattenedSteps: PlanStep[],
-    expectedOutputs: string[]
-  ): string {
-    const lines = [
-      `Derived from linked plan "${plan.ref.slug}" for PREVC phase ${phase} (${getPhaseDefinition(phase).name}).`,
-    ];
-
-    if (matchingPlanPhases.length > 0) {
-      lines.push(`Plan phases: ${matchingPlanPhases.map((planPhase) => planPhase.name).join(', ')}.`);
-    }
-
-    const summaries = matchingPlanPhases
-      .map((planPhase) => planPhase.summary)
-      .filter((summary): summary is string => typeof summary === 'string' && summary.trim().length > 0);
-    if (summaries.length > 0) {
-      lines.push(`Phase objectives: ${summaries.join(' ')}.`);
-    }
-
-    if (flattenedSteps.length > 0) {
-      lines.push(`Key steps: ${flattenedSteps.map((step) => step.description).join('; ')}.`);
-    }
-
-    if (expectedOutputs.length > 0) {
-      lines.push(`Expected outputs: ${expectedOutputs.join(', ')}.`);
-    }
-
-    return lines.join(' ');
   }
 
   private async loadHarnessBinding(): Promise<WorkflowHarnessBinding | null> {
